psd: fix interpretation of layer mask/adjustment layer data lengths
The standard says:
> Size of the data: Check the size and flags to determine what is or is not present. If zero, the following fields are not present

Up to 18 bytes can be read safely in this way. After the 2 bytes for padding
(if size == 20), the rest of the data is conditional on each flag allowing
the data, and the block length allowing the read.

BUG:465148
(cherry picked from commit 291a841ec2bf72459b2c17b9106c7eafc766037c)