paintops: Fix build with GCC
zug::map uses the first parameter to target the object being mapped.
However, member functions always implicitly pass `this` as the first
parameter. This, however, is impossible to do through templating alone.
The function needs to be wrapped in a `std::function`, a `std::bind`,
a lambda, or std::mem_fn. Dmitry prefers the latter as it keeps the
code in line with the Lager/Zug coding patterns.

GCC UCRT64 12.2 enforces this and causes a cascade build failure.

The functionality that Dmitry attempted to use here is actually a C++
proposal, not part of C++17:

https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2017/p0312r1.html

See also: https://devblogs.microsoft.com/cppblog/trip-report-evolution-working-group-at-the-summer-iso-c-standards-meeting-toronto/

Co-authored-by: Dmitry Kazakov <dimula73@gmail.com>
Part-of: <https://invent.kde.org/graphics/krita/-/merge_requests/1787>