Fix hiccups when image has a big filter layer
The hiccups were caused buy the thumbnail generator that tried
to calculate the layer's exact bounds right in the middle of the
stroke. It did that in the GUI stroke, which could delay the
updates up to 200ms, which is quite visible to the user.

The patch moves thumbnail generation into a background stroke.

It uses the same technology as Overview Docker and extracts it in
a form of KisIdleTasksManager and KisIdleTaskStrokeStrategy. In the
following patch I will port OverviewDocker and other places to use
KisIdleTasksManager.

BUG:459517