Implement CI scripts for building an AppImage for Krita
The AppImage is built for 'master', 'krita/*' branches
and all the MRs. The script also runs all the unittest and
generates the report.

This patch also disables coverage reports, because they
would mean that all AppImage packages will be annoteted
with coverage (and slow), which is not something we really
need.

The same applies to ASAN switch: it is disabled, because
Qt should also be built with ASAN to work properly inside
the appimage.