Move ext_qt dependency into its own variable
Since this is a primary dependency and our workflow involves having fork
of Qt which we tweak more or less frequently, having this variable
allows us to disable buildling of Qt (by setting QT_DEP to "") if we
want to install from our fork. Otherwise we have to comment each line
individually everytime we have to build deps, which is a hassle.