Make Transform Tool work on multiple layers
An attempt to let the Transform Tool work on multiple selected layers.
It seems to work okay, but I'm not sure what cases to check for thorough testing, and doubtful about the code.

It works by checking if the selected layers are transformable and then adding them to the processedNodes list returned from KisTransformUtils::fetchNodesList(), which Group Layers use for transforming child layers.
I'm not sure if that is a reasonable thing to do codewise.

There's a feature request for it here: https://krita-artists.org/t/ability-to-transform-multiple-layers-without-having-to-put-them-in-a-group-layer/3396