macos: allow loading mlt plugins from install directory
MLT on macOS assumes it is run from a bundle, however this is not the
case when running krita development envs. We run krita.app from a
bundle, since its needed by the OS, but the bundle is empty, loading
everything from the install directory.

This patch makes mlt look for its plugins and resources in the install
directory if the bundle directories are missing.