Remove automatic "relative mode" activation in the slider spinboxes
Removes automatic "relative mode" activation when the user clicks near
the current value. The "relative mode" can still be activated by
pressing the shift modifier when clicking. This mode consists in
using the dragging distance to change the current value by that amount,
instead of directly jump to the new value where the user clicked.