Fix font-size-adjust calculation from QTextDocument
The previous calculation was wrong, causing text to be stuck in a
specific size regardless of the actual font size setting in the rich
text editor.

When trying to fix it, I found that QFontMetricsF gives bogus values for
xHeight (at least on Windows) so it cannot be used. I ended up
calculating the value by reading the parameters with FreeType. If the
font does not have xHeight, then we just don't write the property.

The value of this property is calculated so that the font size of the
primary font is the same as when this property is not set.

BUG:470782