Fix Palettize filter dialog on HiDPI
On >1x scaling the dither and palette widgets would have cut-off icons.
On >2x scaling the Palettize filter dialog would hang because of
trying to loop by adding the length of a 2x2 dither pattern divided by
the scaling factor (y += 2/2.5 gets rounded to y += 0, infinite loop).
This also applies to the dither widget in the Gradient Map filter.