svgtexttool: Support using SVG wrapping modes
This is a rough attempt at making the current text tool able to handle
some SVG wrapping modes:

**SVG Hard Wrap** uses `white-space:pre` to allow the SVG layout engine
to handle line breaking and `line-height`, instead of relying on the
legacy method of setting `dy` with metrics from `QTextLayout`, which can
be inaccurate in a lot of cases.

**SVG Auto Wrap with Fixed Width** is similar but uses
`white-space:pre-wrap` in combination with `inline-size` to allow for
automatic line breaking to a fixed width. The rich text editor also
displays the text with the same auto-wrap width, though there is no
guarantee that it will produce the same result as the SVG layout.

The pre-existing wrapping mode is named **Legacy Hard Wrap** and is
still the default when creating new text shapes.

This commit does not include a method to set the inline-size. This will
be added separately. (It can be manually set from the SVG.)