Welcome Page: Fix bug where docker size is not preserved
The bug happened because when we booted Krita we called the
saveMainWindowSettings, which resulted in bogus settings being saved. So
we don't call that method when we are booting. But only when we go back
and forth from canvas page.

(cherry picked from commit 953196046c27ff9413b76fad1a3d38582a180ca4)