Set useRichText for newly-imported text to false
If we are importing from an SVG file, it is more likely for text
elements to contain markup that cannot be converted to rich text
losslessly, so it makes more sense to prefer the SVG editor.

This change does not affect new text and any existing text in kra files
created on and after Krita 4.2.0, but existing text from older files may
open the text editor in SVG mode. The user can simply switch back to the
rich text editor.