Fix multichannel filter configuration use of properties
The multichannel filter (inherited by the curves and cross channel filter) uses
member variables to store the curves instead of the QVariant properties of the
KisPropertiesConfiguration. This makes it impossible to read/write the options
through those properties, for example from python side.
This patch just intercepts the functions that set/get properties and transform
the stored KisCubicCurve objects to/from string representation to avoid changing
the mechanics of the whole filter configuration