Fix converting text with multiple alignments back to rich text
If you set different alignment on separate blocks of text, their x
positions in the output SVG will be changed in order for the text-anchor
to work. In this case we need to allow the x to be different to
successfully round-trip it back.

We also need to not apply the top-level block format when reading the
closing tag, because it will overwrite the alignment and other formats
of the previous block. Only apply it on reading the next tag, after
creating the new text block.