text: Add synthesizing italic glyphs
If the user requested italics but the font does not have native italics
variant, 5.1 and earlier will synthesize oblique by shearing the glyphs.
It is unclear what amount of shear Qt used.

Implement the same thing for the new text engine. A shear angle of 14
degrees is chosen to coincide with CSS's defaults. [1] This commit does
not provide an option for disabling this behaviour. If such a switch is
to be implemented in the future, it should probably use the CSS
`font-synthesis-style` property to do so. [2]

CCBUG:465886

[1]: https://drafts.csswg.org/css-fonts/#font-style-prop
[2]: https://drafts.csswg.org/css-fonts/#font-synthesis-style