Pass image dpi to shapes, fix text shape not updating dpi properly
Funnel the image resolution from KisShapeLayer through
ShapeLayerContainerModel to the KoShape instances. This allows KoSvgTextShape
to get the image resolution outside of paintComponent.

The previous approach of calculating the resolution in paintComponent
did not work well because shapes are cloned to be painted in an update
job. This meant the original shape never actually gets the proper dpi
and layout, which causes the shape outline to become out of sync with
the painted shape. This had significant issues for fonts containing
bitmap glyphs, which gives drastically different metrics under different
dpi. (It also meant text shapes always do a relayout whenever they are
painted.)

CCBUG:472566