Make the thumbnails a bit smaller
This also has the beginning of the banner loading code, but I'm
not going to finish that today, so it's commented out.

I've also tried to to put the recent files in a longer block on
the left of the menu columns, but Qt's layout system didn't want
that either.

(cherry picked from commit 704722933baa270cc6da421f4f100be2675e7009)