[string-freeze exception] Fix loading TIFF images with embedded PSD data
In some cases the embedded PSD data may be loaded incorrectly into
Krita, in such cases we should let the user to choose, which form
he/she wants to load.

The patch breaks the string freeze, since it adds a dialog allowing
the user to choose between the two options.

BUG:376950
CC:kimageshop@kde.org
CC:kde-i18n-doc@kde.org