text: Add synthesizing bold glyphs
If the user requested bold but the font does not have native bold
variant, 5.1 and earlier will synthesize bold by making the glyph
thicker.

Implement the same thing for the new text engine. This emboldens the
glyph outline by a bit using `FT_Outline_Embolden`, and adjust the glyph
advance to accommodate for the wider glyph. This implementation is
heavily inspired by Firefox.

For bitmap glyphs, use `FT_Bitmap_Embolden` like `FT_GlyphSlot_Embolden`
does, but we use a lighter strength, and for small fonts size we only
embolden the glyph horizontally. This is similar to what Windows does.
If we embolden both horizontally and vertically, especially for CJK the
glyphs often lose the details and become completely unrecognizable black
blobs.

This commit does not provide an option for disabling it, but we should
implement the `font-synthesis-weight` CSS property to do so in the
future.

BUG:465886