Fix line-height being capped due to naive handling of negative ascent.
Because ascent and descent can be negative or positive depending on the
writing mode, the intial code attempted to regard them somewhat agnostic
until it was time to calculate the precise offset. This however capped
the offset to 0, and made it impossible to have negative line-offsets.

The code now considers negative values appropriately.

BUG:472316