Let baseline adjustments be taken into account for line-height calculation.
To do this, we are now calculating the baseline before applying it inside the
break-lines option. Without this, baseline alignment and shift would overlap
the previous lines, sometimes even flow outside shapes and therefore be not as
useful as they are when taken into account. This may also fix some minor issues
like baseline-shift being interpeted incorrectly as well as better handling
of vertical-lr.