Embed our font libraries into the AppImage
We require quite new font libraries for Krita with experimental features,
so we cannot expect system libraries to be present.

I first tried to use the same fallback approach like we use for
libstdc++, but then remembered that these libraries do not guarantee
backwards compatibility, so we shouldn't try that.

BUG:472710