Enable text tool to draw debug outlines of text shapes in view
This debug function can be enabled by setting the environment variable
`KRITA_DEBUG_TEXTTOOL`. When enabled, selecting a text shape with the
text tool will draw the bounding boxes of the characters in the shape.
The boxes are drawn on the view, allowing the subpixel positioning to be
visualized when zoomed in.

(cherry picked from commit 466d31888d7d4521b50e606beb2d484c61de9181)