Make transformationResolution happen earlier and be a bit better aligned to the SVG 2 spec.
A bit contradictory, becuase the SVG spec wants it to happen later, but we
should have it this early to eventually apply shaping breaks.

Collapsing is now handled by the resolveTransforms function.

This also handles adressability and 'middle' glyphs a little better,
which means we have less problems with ligatures.

BUG:472382
CCBUG:471374