Implement arbitrary run breaks inside raqm to ensure that absolute chunks are broken up.
This only really works for latin, as it misses bidi handling as well
as harbuzz being a little too smart by half and still shaping text.

This patch is a stripped version of this MR:
https://invent.kde.org/graphics/krita/-/merge_requests/1867

See discussion in the MR for the details why the full version has
been rejected in the end.