Make halftone filter work well with python
The generator configurations are cached, so we need to check if the
property name of a property being set represents a property of a
generator configuration, and in that case we must remove the cache entry
that property belongs to, so that the configuration can be regenerated
later. This is an issue mainly when setting the properties directly
(through python for example) instead of using the high level methods.


(cherry picked from commit a741ec529da50e3a21abeaa757c10acf0357304a)