svgtexttool: Remove setting to enable only rich-text or SVG editor
Eventually we shall have on-canvas rich-text editing and the editor
will only allow editing the SVG source. For the time being, we still
keep the rich-text editor for backward compatibility while on-canvas
editing is being implemented, but we can remove the setting to enable
only a specific editor now since it will no longer be useful in the
future, and this allows us to simplify things a bit.