Fix duplicated calligraphy profile save operation.
This fixes a bug where, after entering an invalid profile name, the
user is prompted twice for a new name and two new profiles are
saved. The cause is a recursive call to saveProfileAs(), which is
redundant because the enclosing loop will continue prompting until a
valid name is given.