Fix cross-colorspace bitBlt with channel flags
The channel flags passed to bitBlt should be applied to the
source pixels in source color space. It means that the channels
should be pre-processed with convertChannelToVisualRepresentation()
before being blended into the destination.

Please take it into account that source **color** flags should be
passed to convertChannelToVisualRepresentation, but **alpha** flags
should go to the final blend operation.

BUG:473479