Add alpha-preserve attribute to ORA import/export
On the `<layer>` element, there's now a new attribute
`alpha-preserve="true"` if the layer has alpha preseve enabled. This
allows that property to round-trip when saving and loading an ORA file,
as well as compatibility with other programs that support alpha preserve
on layers, which at the time of writing is Drawpile Programs that don't
support it will just ignore the extra attribute.

Normal with alpha preserve is still saved as `svg:src-atop`, it was
effectively the only blend mode that supported alpha preserve before.
Reading-wise, the code handles both representations.

This also fixes the bug of any blend mode that's not known by the ORA
spec with alpha preserve enabled getting saved as `svg:src-atop` as
well, which effectively clobbered whatever blend mode was set with
Normal. Instead, they now get saved as the proper blend mode with alpha
preserve enabled.