Handle string-index-differences caused by bidi marks or transforms
Bidi insertion causes a difference between the original plaintext
string and the final result. Similarly, text-transforms can sometimes
cause combining marks to be inserted or removed. We should keep track
of which index in the new string maps to which index in the plaintext.

This will make manipulating text on the KoSvgTextShape a bit more
straightforward.