flake: Fix transformation of text path and shape-inside
Shapes referenced by text path and shape-inside should ignore any
transformation of their parents. When the shapes are loaded from SVG,
they have the absolute transformation baked in. In order to get the
transform we need, we invert the transform of a shape's parent and
combine it with the shape's own absolute transform.

CCBUG:472571