text: Draw a hex tofu block for missing glyphs
If for whatever reason we don't have a glyph to render for a particular
character, draw a rectangle containing hex characters showing the
character's Unicode codepoint (like Firefox does). These cases include:

- No suitable glyph in font (replaces the default "tofu" glyph)
- Unsupported glyph format (e.g. SVG glyph)

Note: This does not change the handling of unpaired UTF-16 surrogates;
they are replaced with the glyph for U+FFFD instead.