Fix updates on autokeyframing with onion skins enabled
When we create a new frame, we may change a set of onion skins
that should be displayed on screen. Therefore, we should issue
a set of specific updates for such a case.

The patch adds a set of more granular signals to KisKeyframeChannel,
so that the addition and removal of the keyframes can be tracked
properly.

The patch also fixes the issue with sigRemovingKeyframe() being
emitted **after** the frame was actually removed from the raster
channel.

BUG:474138