[test before 5.2.0] Fix cancelling of the selection when doing two consequtive selections
Back in 2013 we used to switch node properties without undo/redo
information. Hence we needed this notification in
KisImage::nodeChanged(). Now we use undo+strokes for the switch,
so requestStrokeEnd() is emitted explicitly by the stroke.

This change looks a bit scary for 5.2.0, because the change is in
a rather deep place, so we should test it well before backporting.

BUG:474525