Fix external ffmpeg not working on Linux due to clash in LD_LIBRARY_PATH
We embed a dynamically linked ffmpeg into our appimage and add that
into the environment with a LD_LIBRARY_PATH variable. If the user
tries to use a system-provided dynamically-linked ffmped, then there
will be a clash and ffmpeg will crash.

Hence we should detect if a non-embedded ffmpeg is running and remove
the variable from its environment.

BUG:473603