Fix SEGFAULT in KoColorSet unit tests.
The cause of the SEGFAULT is the deletion of a dangling pointer in the
destructor for KoColorProfileStorage::Private. The offending pointer
should actually be removed from the 'duplicates' list by a call to
removeProfile() since it will probably be freed afterwards.

This change also removes a redundant conditional in removeProfile().