Use 'OBJECT' instead of 'STATIC' linking to transform tool plugin
Otherwise the registrar is not pulled into the plugin, so ld just
drops it completely.

The problem is that static linking assumes that unused symbols are
just dropped from the final object file, and our registrar objects
are not used anywhere outside the static library itself (obviously).

There are two solutions to this problem:

1) Use OBJECT linking instead of STATIC (might cause troubles on macOS,
   needs testing)

2) Make sure that all registrar objects are defined in the actual
   .so library of the plugin, not in its static part.