WIP: Refactor transform masks not to discard keyframe channels on every change
The patch fixes the weird design of the transform masks. Before the patch
KisAnimatedTransformMaskParameters were replaced on every transform mask
change, which also replaced the keyframe channels. Which means that
every undo operation would just **delete** the keyframe channels
linked inside the node itself.

In the new design KisAnimatedTransformMaskParameters is never replaced
(it is actually planned to rename it into KisTransformMaskParametersHolder
or something like that). The holder just stores keyframe channels and
the params and just returnes the proper paraps when requested (by fetching
some information from the keyframe channels).

BUG:475385