Fix Brush Editor sizing issues
This is ~~three~~ four commits. 
The first saves the brush editor's window geometry in the config.
The third removes a function that doesn't appear to be used. The fourth fixes the initial, unconfigured sizes.

The second one reverts the brush editor's sidebar toggle behavior:
When the brush editor was changed to use a QSplitter, it became that 
the presets and scratchpad sidebars shared the width of the window with
the center editor widget. This led to showing/hiding them causing the
center widget to be either cut off or stretched.
This commit changes it back to the previous behavior of
changing the width of the window instead, leaving the center untouched.

The window resizing when doing this is unfortunately a bit flickery visually, as before.
I don't know whether that can be fixed. (Might be OS-dependent?)
There is a difference from 5.1.x; before when toggling the presets,
the window would stay in place and the widgets move over.
Now the window moves and the widgets stay in place.

With these ~~two~~ three commits building on the ability to save the splitter sizes (MR!1878),
the user has control over the sizes of the brush editor and they will be saved.
~~This doesn't however fix the issue of the initial, unconfigured sizes being
cut off.~~ The default size will be wide enough that the widgets won't be cut off, but can be configured smaller to save space.