Add brush tip rotate actions
Users requested a shortcut to modify brush tip rotation while staying
on canvas. This change adds 4 new actions, rotate clockwise,
rotate counter-clockwise, and at two granularities coarse and precise.

Implemented for brush-based paintops and for MyPaint. Other brush
engines don't seem to benefit from the tip rotation too much.

The rotation paintop is defined as an increment + direction (cw/ccw)
because the rotation direction as well as the unit (radians/degrees)
may depend on the brush engine.

The path is a feature request from this thread: 
https://krita-artists.org/t/shortcut-to-rotate-brush/43675