Add update-ahead feature to outline updates
Due to a compression for widget updates in Krita and Qt, the brush
outline may change while the updates are being passed through the
pipeline.

We can either pass the outline alongside the update itself (which is
a bit complicated) or just do a simple update-ahead. This patch
implements the latter approach. It just calculated an approximation
of the next outline position and extends the update rect to that.

BUG:476300