Improve IME decoration after studying Qt's platform integration.
Basically, when using Japanese IME on Windows and MacOS, it is
possible to 'select' certain parts of the input string. When this
is happening, the underline should be thicker, but because Qt doesn't
allow setting this information, instead the background gets set.

For that purpose we're making the underline thicker when the background
is set.