Add BrushRotation angle selector to KisPaintopBox
This PR adds a selectable "slider" widget with an angle selector for the recently added brush rotation. This should be the last entry in the brush rotation saga :) Implements one of the feature request from [this post](https://krita-artists.org/t/more-brush-option-slider-options-and-buttons-to-turn-off-pressure-sensitivity/33647).

The new control in the top bar:

![image](/uploads/daaaf2b2aaca49b18382622c978ce38c/image.png)

Demo video (this uses angle snapping of 10 degrees, but it is now changed to the standard 15 degrees):

https://www.youtube.com/watch?v=ztdXgNjYhgg

Test Plan
---------

Use the dropdowns on one of the sliders in the top bar to access the widget and make use of it. Pin it to the bar with the + icon and test more. The video shows just that.

Formalities Checklist
--------------------- 

- [x] I confirmed this builds.
- [x] I confirmed Krita ran and the relevant functions work.
- [ ] I tested the relevant unit tests and can confirm they are not broken. (If not possible, don't hesitate to ask for help!)
- [x] I made sure my commits build individually and have good descriptions as per [KDE guidelines](https://community.kde.org/Policies/Commit_Policy).
- [x] I made sure my code conforms to the standards set in the HACKING file.
- [x] I can confirm the code is licensed and attributed appropriately, and that unattributed code is mine, as per [KDE Licensing Policy](https://community.kde.org/Policies/Licensing_Policy).

_**Reminder: the reviewer is responsible for merging the patch, this is to ensure at the least two people can build the patch. In case a patch breaks the build, both the author and the reviewer should be contacted to fix the build.**_
_**If this is not possible, the commits shall be reverted, and a notification with the reasoning and any relevant logs shall be sent to the mailing list, kimageshop@kde.org.**_