Fix an offset when opening files with animated transform masks
Our curves should be centered againt the rotation center.

The current implementation still has a bug that the curve value
is **not** the position of the rotation center, but (position -
rotationCenterOffset), which is a completely nonsensial value.
But since it is baked into .kra format, we should keep it this
way for now.

BUG:476317