Text: Implement cross layer shape selection for the text tool.
A request from artists, this helps navigate complex compositions
like comics.

This adds two new (and one helper) functions to KisToolUtils:

shapeHoverInfoCrossLayer will retrieve information useful for hovering.
selectShapeCrossLayer will select the layer and shape under the cursor.

The shape itself is not returned by these functions as it is possible
to do silly things with said shape, like selecting it on a different
layer's shape selection.

The function is set up so shape edit and selection could use it, but
not implemented yet as these tools can select multiple shapes.