Add recorder thread activity to REC status bar
Besides the REC label and the red recording symbol, the REC status bar
displays now also the maximum number of threads, which can be
theoretically used for recording.
Unused threads are displayed in a dark gray. Threads in idle state have
a light green and active threads are shown in a dark green. If the
number of used threads exceed the IdealRecordThreadCount, they are
displayed in orange. And if they exceed the MaxRecordThreadCount, they
are displayed in red.

Adapt the behaviour of the red REC symbol activity ascertainment.
The activity can now be determined by active recording threads.
Add a QTimer for the Rec symbol activity timeout. The timeout interval
is set depending on the capture interval (slightly bigger, to avoid
flickering), but not smaller than 305ms and
not greater than 2005ms. The timer is started/restarted with every
increasing change of the used recording threads. On timeout the
recording symbol is repainted in its inactive color.
Remove unused pause flag and pausedChanged Signal.