Rework Recording
Change RecorderWriter to RecorderWriterManager, which is now just a
QObject. It keeps a QThread member, which times out continuously
depending on the capture interval. But the thread event handler of this
QThread member doesn't capture the current image anymore. Instead, it
triggers one of the new RecorderWriter instances, which are handled each
by a separated thread. This refactoring allows us to capture
several images in parallel, even if a single capturing step takes more
time than the capture interval.

Add sanity checks to setCanvas() ans setup().
Methods can now also be called during an active recording

Rename signal finish to stopped

The signals started and stopped (previously finished) where emitted by
QThread, but due to the fact, that RecorderWriterManager is now a
QObject we have to emit them explicitly