Text: Implement initial pos function.
Turns out child x and y transforms do affect an inline size, but only
the start position. This reverts the change that made the first pos on the text
the root position, and now tries to find a proper start pos during text layout.

This also improves the default cursor for text-on-path and in-shape.