svgtexttool: Do not replace canvas resource foreground color
This was added as part of 1913d805b8 to fix bug 443793, but this
particular change is not necessary. The svgtexttool does not actually
have any code wired to change the text color from the foreground color,
so setting the foreground color at all is simply misleading.

This may be properly reimplemented as part of rich text editing support
of the text tool rewrite later.