defaulttool: Delay tool switch request to the Qt event loop
When double-clicking or pressing Enter on a shape to edit it, we avoid
an issue that causes events to be erroneously resent to `KisToolProxy`:

The original event is handled by `KisShortcutMatcher` from
`KisInputManager`, which then goes through `KisToolInvocationAction` and
`KisToolProxy` to eventually reach `DefaultTool`. When the tool calls
`KoToolManager::switchToolRequested` directly, it eventually calls
`KisShortcutMatcher::lostFocusEvent` and
`KisShortcutMatcher::toolHasBeenActivated`, both of which triggers the
recursion guard. `KisInputManager` then thinks the event hasn't been
handled and calls `processUnhandledEvent`, which sends the event to
`KisToolProxy` for a second time. Observable effects include: double-
clicking a text shape causes all text to be selected (by
`mouseTripleClickEvent`), or pressing Enter on a text shape causes an
unwanted newline to be inserted immediately.