svgtexttool: Select all text when first selecting a text shape
This recovers the previously accidental behaviour that, when switching
to the text tool from DefaultTool by double-clicking on a text shape all
its text will be selected. In addition this also now happens when
pressing Enter to edit a text shape, and also when creating a new text
shape by dragging.