Fix invalid XML in bundle metadata (meta.xml) file.
The "meta" and "dc" namespaces in the metadata file should be declared
as specified in the OpenDocument schema. QtXml doesn't complain about
this currently, but some parsers fail when namespaces are undeclared
(e.g. python's minidom).

The <cd:creator> tag in the bundle metadata file (meta.xml) should
actually be <dc:creator>. This appears to be the result of a typo in
7bf66b17b7a, since the correct form appears prior.