Fix inconsistent undo/redo of select/deselect/reselect actions
Before the patch, the global selection actions operated with a
raw selection object, not with a selection mask. It caused issues,
when the user manually converted/moved the global selection mask
into another position.

The patch changes this behavior and makes the selection commands
operate with the entire selection mask, which makes operations
consistent.

BUG:478519,478520