Implement a gap closing contiguous fill
This patch modifies the behavior of the scanline fill algorithm
to cooperatively work with a gap closing flood fill. This method
can make use of a "gap distance map" to identify the lineart gaps
and stop the fill before spilling.

The core distance search algorithm was adopted from MyPaint.

Modified tools:
- contiguous (bucket) fill - can now specify a gap size up to 32 px
- contiguous selection tool - same, but for creating a selection
  that stops on gaps

Excluded:
- enclose and fill tool - it is already rather complex, but uses
  a similar code path underneath. Maybe revisit at a different time.

Also adds regression tests to check the fill's behavior on a small
resolution test image.

I believe there are still some opportunities for optimization,
but this initial version should be good enough for a lot of use cases.