Add colorspace lock to Specific Color Selector
Previously, unless the CS combobox was changed, the internal colorspace
would match the layer, without changing the UI. When clicking the
RGB/HSV button in a nonRGB layer, it would then hide itself
despite the UI still being in RGB.

Now there's a lock to indicate whether the colorspace should use the
current layer or a colorspace specified by the user in the combobox,
and the UI behaves accordingly.

Also, simplifies KisSpecificColorSelectorWidget::updateHsvSelector().

BUG:478521