Fix a crash when quickly undo/redo new-from-visible+visibility-change
This patch introduce multiple changes to fix the original bug:

1) KisSynchronizedConnection no longer behaves as if it were
an auto-connection. Instead, **all** signals are now linearized
via the qApp events queue. This fixes the reordering issue when two
actions are undone/redone too quickly and one action is a legacy
action that executes in the GUI thread (e.g. visibility change) and
the other action is a "processing" action that executes in a worker
thread (e.g. new-from-visible). "Auto"-style connection would reorder
the actions on the basis of their execution thread.

2) Unittests still explicitly activate the "auto"-style connection,
because they don't have KisApplication and the corresponding handler
for KisSynchronizedConnectionEvent.

3) Fixes a bug in KisDummiesFacadeBase that caused double addition
of the dummies on loading of the document. It could happen that
two setImage() calls were called in the GUI thread **before** the
addDummyImpl() had a chance to execute.

BUG:478225