tiff: Prefer the ADOBE version of the DEFLATE tag
The `COMPRESSION_DEFLATE` was first introduced as a private tag, and
then as been standardized by Adobe as the `COMPRESSION_ADOBE_DEFLATE`.
The former is now considered as non-standard and deprecated, libtiff
issues a warning when encoding an image with it.

This patch replaces all occurences of the old tag by the new one and
has been generated with the following command:
```
find . -not -path '*/\.git/*' -type f -exec \
sed -i 's/COMPRESSION_DEFLATE/COMPRESSION_ADOBE_DEFLATE/g' '{}' \;
```