tiff: Refactor duplicated writer code in a base class
Both normal TIFF and PSD writers are based on the same code. It's pretty
clear than the latter was created by copying the code source of the
former. So let's merge these two codes before they diverge even more.
This commit only includes the part with only whitespace changes. There
is still a lot more to factorize but that will need more care.