Remove unnecessary locking code from several assistants
Except for perspective and 2-point perspective, there is only ever one possible
curve per assistant that goes through the starting point, so there is no need
to wait for movement.
And they did not only do that on stroke begin, but every time the cursor gets
near the stroke start again.

Inhibiting the movement is now done by the code deciding for the best
assistant, because it is not necessary when no other assistant offers an
alternative curve nearby.