Put docker panel in Popup Palette
This panel borrows the widget of the selected docker
to display it in the Popup Palette, while the panel is open.
If the docker itself is open, it shows a message explaining
where the widget has gone.

The On-Canvas Brush Editor formerly in this panel is
converted into a docker, and chosen by default.
The list of dockers shown in the combobox is configurable.