Fix Overview docker in Popup Palette panel
Refactor the handling of events to not be on the Overview QDockWidget,
but on its owned widget, so it can be borrowed without issue.
This fixes the issue of the widget not resizing when borrowed/returned,
and lets the events handling (hovering the controls to show them,
for example) be borrowed.