Fix crash when pasting font family in TextEditor
Problem:

Pasting a new font famyly and hitting Save or Discard while the
FontFamilyComboBox is still in editing focus crashes.
After pasting font family, the font is also not applied until you
press enter or change focus.

Solution:

Write the validator of the font familyies.
Validator is triggered whenever the text is changed (including pasting),
and if font famyly is correct, the font changes immediately.
So, there are no conditions that provoke a crash.

BUG:484066