Fix grids glitches when subdivision style is solid
Bug has been introduced with commit 764719791a4ce6887aedc9d92777f3c4a03198c6 (to fix bug 461912)

From Qt's documentation, setDashOffset() method implicitly converts the style of the pen to Qt::CustomDashLine.
When current line style is solid, it seems the custom dash line is not soid anymore and this generate glitches.

To avoid the problem, need to to check if style is solid or not before updating dash offset (update only if not solid).

In addition, the original commit only 'fixed' dash offset problem for subdivision but not for main grid division; this commit fix both (main & subdivision) grids lines.

BUG:484889